refactor: removed unused features in configuration
